in p5 change width to canvasWidth
change name of width and height to be less generic
remove 'virtual' from various IGraphics sub-class interfaces

+++ primary/Network.hpp
@@ -182,7 +182,7 @@ public:
         return true;
     }
 
-    int numberOfAddresses() {return fAddresses.size();}
+    size_t numberOfAddresses() {return fAddresses.size();}
 
     IPAddress * getAddress(const int idx = 0)
     {
